% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateKnn2.R
\name{calibrateKnn2}
\alias{calibrateKnn2}
\title{calib2}
\usage{
calibrateKnn2(Xr, Xm, dr = NULL, dm = NULL, knn, tol = 0.001, max_iter = 2000)
}
\arguments{
\item{Xr}{a matrix without NA values. The rows correspond to respondent units.}

\item{Xm}{a matrix with at least one NA values on each of its rows. The rows correspond to nonrespondent units.}

\item{dr}{a vector containing the sampling weights of the respondent units in \code{Xr}. If NULL (default), all sampling weights are equal to 1.}

\item{dm}{a vector containing the sampling weights of the nonrespondent units in \code{Xm}. If NULL (default), all sampling weights are equal to 1.}

\item{knn}{a matrix that is returned by the function \code{\link{indKnn}}.}

\item{tol}{a tolerance parameter. Default value is 1e-3.}

\item{max_iter}{the maximum number of iterations to consider convergence.}

\item{k}{the number of neighbors that are candidate to impute each nonrespondent. If NULL (default), the smaller k
that makes possible to resolve calibration equations is chosen.}
}
\value{
Returns a list including:

\code{knn_final} a matrix that is returned by the function \code{\link{indKnn}}.
If \code{knn} is not null and is to small to converge, \code{knn_final} will consider a larger number of neighbors. Otherwise, \code{knn_final} will be equal to \code{knn}.

\code{psi} a matrix with the same length as \code{knn_final} containing imputation probabilities for each \code{k} neighbors.
}
\description{
Compute imputation probabilities matrix satisfying calibration equations.
Only the \code{k}-nearest respondent units of each nonrespondent have non-zero imputation probabilities.
}
\details{
The main idea is to have imputation probabilities satisfying some calibration constraints for all variables simultaneously
and summing to one for each nonrespondent. The method is based on four requirements:
\enumerate
      \item The imputed values is selected among the completely observed units: a donor imputation method is used.
      \item Only one donor is selected per nonrespondent: all missing values of a unit should be imputed by the same donor.
      \item The donors is selected among the \code{k} nearest neighbors of units with missing values.
      \item If the observed values of the nonrespondents were imputed, the total estimator of each variable will remain unchanged.
\enumerate
See the package's vignette to a complete description of the calibration constraints.
}
\examples{

}
\seealso{
\link{indKnn}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
