% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeImput.R
\name{cubeImput}
\alias{cubeImput}
\title{Donor imputation using the cube method}
\usage{
cubeImput(Xr, Xm, knn, psi)
}
\arguments{
\item{Xr}{a matrix without NA values. The rows correspond to respondent units.}

\item{Xm}{a matrix with at least one NA values on each of its rows. The rows correspond to nonrespondent units.}

\item{knn}{a matrix containing in rows the nonrespondent units and in columns the decreasing rank
of the \code{K}-Nearest Neighbour among the respondents units.}

\item{psi}{a matrix with the same dimension of \code{KNN} contains imputation probabilities of the
\code{K}-Nearest Neighbor of the nonrespondent units.}
}
\value{
the dataframe of nonrespondents with NA values imputed.
}
\description{
Impute missing values with one of the nearest respondents using the cube method.
}
\examples{
df         <- matrix(c(1,2,1,3,2,
                       NA,3,4,5,2,
                       1,3,NA,NA,3,
                       3,2,5,7,8,
                       5,3,4,2,2,
                       6,NA,2,2,1,
                       4,2,4,4,3,
                       2,5,6,3,7,
                       1,4,2,3,1,
                       NA,NA,NA,NA,2), nrow = 10, ncol = 5, byrow = TRUE)
res        <- indKnn(X = df,  k=5)
KNN        <- res$KNN
res1       <- calibrateKnn(X = df, d = NULL, KNN = KNN, PSI.init = NULL, k = NULL, max.iter = 1000, tolerance = 0.01)
k_opt      <- res3$k_opt
PSI        <- res3$PSI
res_imput  <- cubeImput(X = df, PSI = PSI, KNN = KNN[,1:k.opt])
df_new     <- res_imput$X_imput # df with NA values imputed
df_new

}
\references{
Deville, J. C. and Tille, Y. (2004). Efficient balanced sampling: the cube method. Biometrika, 91(4), 893-912.
\emph{Biometrika}, 91(4), 893-912

Chauvet, G. and Tille, Y. (2006). A fast algorithm for balanced sampling.
\emph{Computational Statistics}, 21(1), 53-62.
}
\seealso{
\link{calibrate.knn} and \link[BalancedSampling:flightphase]{BalancedSampling::flightphase}
}
\author{
Esther Eustache, \email{esther.eustache@unine.ch}
}
