% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministicImput.R
\name{deterministicImput}
\alias{deterministicImput}
\title{Linear imputation of Swiss cheese nonresponse}
\usage{
deterministicImput(X, d = NULL, k = NULL, tol = 0.001, max_iter = 2000)
}
\arguments{
\item{X}{a matrix with NA values. The rows correspond to the units.}

\item{d}{a vector containing the sampling weights of the units. If NULL (default), all sampling weights are equal to 1.}

\item{k}{the number of neighbors considered to impute each nonrespondent. If NULL (default), the smallest \code{k}
allowing a solution to the calibration equations will be chosen.}

\item{tol}{a tolerance parameter. Default value is 1e-3.}

\item{max_iter}{the maximum number of iterations to consider convergence.}
}
\value{
Returns a list including:

\code{X_new} the imputed matrix of \code{X}.

\code{k} the number of neighbors taken into account.
}
\description{
Impute missing values by means of a deterministic approach using a matrix of imputation probabilities.
}
\details{
First, the \code{k} nearest neighbors of each nonrespondent are chosen using the Euclidean distance in the function \code{\link{indKnn}}.
Next, imputation probabilities are computed for the nearest neighbors of each nonrespondent.
The imputation probabilities satisfy some calibration constraints for all variables simultaneously.
The matrix of imputation probabilities is computed using the function \code{\link{calibrateKnn}}.
Then, each missing value is imputed by the sum of the multiplication of the imputation probabilities and the corresponding observed values.
}
\examples{
Xr  <- rbind(c(0.1,0.3,0.4,0.1), c(0.1,0.3,0.2,0.1), c(0.1,0.2,0.3,0.1),
             c(0.2,0.3,0.2,0.3), c(0.1,0.1,0.2,0.1))
Xm  <- rbind(c(NA,0.1,NA,0.1), c(0.1,NA,0.2,NA))
X   <- rbind(Xr,Xm)
deterministicImput(X)

}
\seealso{
\code{\link{indKnn}} and \code{\link{calibrateKnn}}
}
