% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotdeckImput.R
\name{hotDeckImput}
\alias{hotDeckImput}
\title{Random hot deck imputation of Swiss cheese nonresponse}
\usage{
hotDeckImput(X, k = NULL)
}
\arguments{
\item{X}{a matrix with NA values. Rows correspond to units.}

\item{k}{the number of neighbors considered to impute each nonrespondent. If NULL (default), the minimum between
30 will be chosen.}
}
\value{
the new dataframe with NA values imputed.
}
\description{
Impute each nonrespondent unit by selecting randomly a donor among its nearest respondents in term of Euclidean distance
(see (Andridge and Little, 2004)).
}
\examples{
 #A faire

}
\seealso{
\link{indKnn}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
