% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputMatrixKnn.R
\name{imputMatrixKnn}
\alias{imputMatrixKnn}
\title{Compute an imputation matrix}
\usage{
imputMatrixKnn(X, psi, knn)
}
\arguments{
\item{X}{a matrix with NA values. The rows correspond to the units.}

\item{psi}{a matrix of imputation probabilities returned by the function \code{\link{calibrateKnn}}.}

\item{knn}{a submatrix of the matrix returned by the function \code{\link{indKnn}}, with the \code{k} columns corresponding to the nearest neighbors.}
}
\value{
The imputation matrix with the same size as matrix \code{\link{knn}} composed of 0s and 1s specifying which respondent will imput each nonrespondent.
}
\description{
Compute an imputation matrix using the cube method and a matrix of imputation probabilities.
Only the \code{k}-nearest respondent of each nonrespondent have non-zero imputation probabilities.
}
\details{
The function \code{\link[StratifiedSampling::stratifiedcube]{StratifiedSampling::stratifiedcube}} from the package \code{StratifiedSampling} is used to compute the imputation matrix to select donors.
This function uses the cube method (see (Deville and Tille, 2004)).
}
\examples{
Xr  <- rbind(c(0.1,0.3,0.4,0.1), c(0.1,0.3,0.2,0.1), c(0.1,0.2,0.3,0.1),
             c(0.2,0.3,0.2,0.3), c(0.1,0.1,0.2,0.1), c(0.3,0.1,0.1,0.1))
Xm  <- rbind(c(NA,0.1,NA,0.1), c(0.1,NA,0.2,NA))
X   <- rbind(Xr,Xm)
knn <- indKnn(Xr,Xm)[,1:5] #--the k = 5 nearest neighbors
psi <- calibrateKnn(Xr, Xm, knn = knn)
imputMatrixKnn(X, psi, knn)

}
\references{
Deville, J. C. and Tille, Y. (2004). Efficient balanced sampling: the cube method. Biometrika, 91(4), 893-912.
\emph{Biometrika}, 91(4), 893-912
}
\seealso{
\code{\link[StratifiedSampling::stratifiedcube]{StratifiedSampling::stratifiedcube}}, \code{\link{calibrateKnn}}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
