% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indKnn2.R
\name{indKnn2}
\alias{indKnn2}
\title{K-nearest neighbors of nonrespondent units}
\usage{
indKnn2(Xr, Xm)
}
\arguments{
\item{Xr}{a matrix without NA values. Rows correspond to respondent units.}

\item{Xm}{a matrix with some NA values. Rows correspond to nonrespondent units.}

\item{k}{the number of neighbors that are candidate to impute each nonrespondent. If NULL (default), the smaller k
that makes possible to resolve calibration equations is chosen.}
}
\value{
a matrix containing in columns the nonrespondent units and in rows the k-nearest respondent units for each nonrespondent the decreasing rank
in decreasing rank.
}
\description{
Find the k-nearest neighbors of nonrespondent units among respondent units, in terms of Euclidean distance.
The function \code{\link[FNN:knn]{knn}} from the package \code{KNN} is used.
}
\examples{


}
\seealso{
\code{\link[FNN:knn]{FNN::knn}}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
