% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearImput.R
\name{linearImput}
\alias{linearImput}
\title{Linear imputation of Swiss cheese nonresponse}
\usage{
linearImput(X, d = NULL, k = NULL, tol = 0.001, max_iter = 2000)
}
\arguments{
\item{X}{a matrix with NA values. Rows correspond to units.}

\item{d}{a vector containing the sampling weights of units. If NULL (default), all sampling weights are equal to 1.}

\item{k}{the number of neighbors considered to impute each nonrespondent. If NULL (default), the smaller k
that makes possible to satisfying calibration equations will be chosen.}

\item{tol}{a tolerance parameter. Default value is 1e-3.}

\item{max_iter}{the maximum number of iterations to consider convergence.}
}
\value{
Returns a list including:

\code{X_new} the imputed matrix of \code{X}.

\code{k} the number of neighbors taken into account.
}
\description{
Impute missing values using the linear imputation.
}
\examples{


}
\seealso{
\link{indKnn} and \link{calibrateKnn}
}
