% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpSwissCheeseImput.R
\name{lpSwissCheeseImput}
\alias{lpSwissCheeseImput}
\title{Donor imputation of Swiss cheese nonresponse using the cube method}
\usage{
lpSwissCheeseImput(X, d = NULL, w = NULL, k = 5)
}
\arguments{
\item{X}{a matrix with NA values. The rows correspond to the units.}

\item{d}{a vector containing the sampling weights of the units. If NULL (default), all sampling weights are equal to 1.}

\item{k}{}
}
\value{
the imputed matrix of \code{X}.
}
\description{
Impute missing values that have no particular pattern by using a donor imputation method.
It extends the balanced \code{k}-nearest neighbor imputation (Hasler and Tille, 2016) to the treatment of the Swiss cheese nonresponse.
}
\examples{
Xr  <- rbind(c(0.1,0.3,0.4,0.1), c(0.1,0.3,0.2,0.1), c(0.1,0.2,0.3,0.1),
             c(0.2,0.3,0.2,0.3), c(0.1,0.1,0.2,0.1))
Xm  <- rbind(c(NA,0.1,NA,0.1), c(0.1,NA,0.2,NA))
X   <- rbind(Xr,Xm)
lpSwissCheeseImput(X)

}
\seealso{
\code{\link[StratifiedSampling:stratifiedcube]{StratifiedSampling::stratifiedcube}}, \code{\link{indKnn}}, \code{\link{calibrateKnn}}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
