% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateKnn.R
\name{calibrateKnn}
\alias{calibrateKnn}
\title{Compute imputation probabilities matrix}
\usage{
calibrateKnn(Xr, Xm, dr = NULL, dm = NULL, knn, tol = 0.001, max_iter = 2000)
}
\arguments{
\item{Xr}{a matrix without NA values. The rows correspond to the respondent units.}

\item{Xm}{a matrix with at least one NA value on each of its rows. The rows correspond to nonrespondent units.}

\item{dr}{a vector containing the sampling weights of the respondent units in \code{Xr}. If NULL (default), all sampling weights are equal to 1.}

\item{dm}{a vector containing the sampling weights of the nonrespondent units in \code{Xm}. If NULL (default), all sampling weights are equal to 1.}

\item{knn}{a matrix that is returned by the function \code{\link{indKnn}}.}

\item{tol}{a tolerance parameter. Default value is 1e-3.}

\item{max_iter}{the maximum number of iterations to consider convergence.}
}
\value{
a matrix with the same size as \code{knn} containing imputation probabilities for each neighbor. If NULL, the constraints of calibration cannot be satisfied.
}
\description{
Compute imputation probabilities matrix satisfying calibration equations.
Only the \code{k}-nearest respondent units of each nonrespondent have non-zero imputation probabilities.
}
\details{
The main idea is to have imputation probabilities satisfying some calibration constraints for all variables simultaneously
and summing to one for each nonrespondent. The method is based on four requirements:
\itemize{
 \item{The imputed values is selected among the completely observed units: a donor imputation method is used.}
 \item{Only one donor is selected per nonrespondent: all missing values of a unit should be imputed by the same donor.}
 \item{The donors is selected among the \code{k} nearest neighbors of units with missing values.}
 \item{If the observed values of the nonrespondents were imputed, the total estimator of each variable will remain unchanged.}
}
See article ... on Arxiv for a complete description of the calibration constraints.
}
\examples{
Xr  <- rbind(c(0.1,0.3,0.4,0.1), c(0.1,0.3,0.2,0.1), c(0.1,0.2,0.3,0.1),
             c(0.2,0.3,0.2,0.3), c(0.1,0.1,0.2,0.1))
Xm  <- rbind(c(NA,0.1,NA,0.1), c(0.1,NA,0.2,NA))
knn <- indKnn(Xr,Xm)
calibrateKnn(Xr,Xm, knn = knn)

}
\seealso{
\code{\link{indKnn}}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
