% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateKnn.R
\name{calibrateKnn}
\alias{calibrateKnn}
\title{Compute a matrix of imputation probabilities}
\usage{
calibrateKnn(Xr, Xm, dr = NULL, dm = NULL, knn, tol = 0.001, max_iter = 2000)
}
\arguments{
\item{Xr}{a matrix without NA values. The rows correspond to the responding units.}

\item{Xm}{a matrix with at least one NA value in each row. The rows correspond to the nonresponding units.}

\item{dr}{a vector containing the sampling weights of the responding units from \code{Xr}. If NULL (default), all sampling weights are equal to 1.}

\item{dm}{a vector containing the sampling weights of the nonresponding units from \code{Xm}. If NULL (default), all sampling weights are equal to 1.}

\item{knn}{a submatrix of the matrix returned by the function \code{\link{indKnn}}, with \code{k} columns corresponding to the nearest neighbors.}

\item{tol}{a tolerance parameter. Default value is 1e-3.}

\item{max_iter}{the maximum number of iterations to consider convergence.}
}
\value{
a matrix of the same size as \code{knn} containing imputation probabilities associated with each neighbor. If NULL, the constraints of calibration cannot be satisfied.
}
\description{
Compute a matrix of imputation probabilities satisfying some calibration equations.
Only the \code{k}-nearest responding units of each nonrespondent have non-zero imputation probabilities.
}
\details{
The main idea is to obtain imputation probabilities satisfying some calibration constraints for all variables simultaneously
and summing to one for each nonrespondent. The method is based on four requirements:
\itemize{
 \item{The imputed values are selected among the completely observed units: a donor imputation method is used.}
 \item{Only one donor is selected per nonrespondent: all missing values of a unit should be imputed by the same donor.}
 \item{The donors are selected among the \code{k} nearest neighbors of units with missing values.}
 \item{If the observed values of the nonrespondents were imputed, the total estimator of each variable will remain unchanged.}
}
See article XXXXXXX on Arxiv for a complete description of the calibration constraints.
}
\examples{
Xr  <- rbind(c(0.1,0.3,0.4,0.1), c(0.1,0.3,0.2,0.1), c(0.1,0.2,0.3,0.1),
             c(0.2,0.3,0.2,0.3), c(0.1,0.1,0.2,0.1))
Xm  <- rbind(c(NA,0.1,NA,0.1), c(0.1,NA,0.2,NA))
knn <- indKnn(Xr,Xm)[,1:5] #--the k = 5 nearest neighbors
calibrateKnn(Xr,Xm, knn = knn)

}
\seealso{
\code{\link{indKnn}}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
