% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministicImput.R
\name{deterministicImput}
\alias{deterministicImput}
\title{Linear imputation of Swiss cheese nonresponse}
\usage{
deterministicImput(X, d = NULL, k = 5, w = NULL)
}
\arguments{
\item{X}{a matrix with NA values. The rows correspond to the units.}

\item{d}{a vector containing the sampling weights of the units. If NULL (default), all sampling weights are equal to 1.}
}
\value{
the imputed matrix of \code{X}.
}
\description{
Impute missing values by means of a deterministic approach using a matrix of imputation probabilities.
}
\details{
First, the \code{k} nearest neighbors of each nonrespondent are chosen using the Euclidean distance in the function \code{\link{indKnn}}.
Next, imputation probabilities are computed for the nearest neighbors of each nonrespondent.
The imputation probabilities satisfy some calibration constraints for all variables simultaneously.
The matrix of imputation probabilities is computed using the function \code{\link{calibrateKnn}}.
Then, each missing value is imputed by the sum of the multiplication of the imputation probabilities and the corresponding observed values.
}
\examples{
Xr  <- rbind(c(0.1,0.3,0.4,0.1), c(0.1,0.3,0.2,0.1), c(0.1,0.2,0.3,0.1),
             c(0.2,0.3,0.2,0.3), c(0.1,0.1,0.2,0.1))
Xm  <- rbind(c(NA,0.1,NA,0.1), c(0.1,NA,0.2,NA))
X   <- rbind(Xr,Xm)
deterministicImput(X)

}
