% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotdeckImput.R
\name{hotDeckImput}
\alias{hotDeckImput}
\title{Random hot deck imputation of Swiss cheese nonresponse}
\usage{
hotDeckImput(X, k = NULL)
}
\arguments{
\item{X}{a matrix with NA values. The rows correspond to the units.}

\item{k}{the number of neighbors considered to impute each nonrespondent. If NULL (default), all the respondents will be considered.}
}
\value{
the new dataframe with NA values imputed.
}
\description{
Impute each nonrespondent unit by selecting randomly a donor among its nearest respondents in term of Euclidean distance.
}
\details{
First, the \code{k} nearest neighbors of each nonrespondent are chosen in terms of Euclidean distance using function \code{\link{indKnn}}.
Next, a donor is selected randomly among the neighbors of each nonrespondent in order to impute the missing values.
}
\examples{
Xr  <- rbind(c(0.1,0.3,0.4,0.1), c(0.1,0.3,0.2,0.1), c(0.1,0.2,0.3,0.1),
             c(0.2,0.3,0.2,0.3), c(0.1,0.1,0.2,0.1))
Xm  <- rbind(c(NA,0.1,NA,0.1), c(0.1,NA,0.2,NA))
X <- rbind(Xr,Xm)
hotDeckImput(X)

}
\references{
Rebecca R. Andridge, Roderick J. A. Little (2010). A Review of Hot Deck Imputation for Survey Non-response. International Statistical Review, Vol. 78, Issue 1, pp. 40-64.
}
\seealso{
\code{\link{indKnn}}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
