% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indKnn.R
\name{indKnn}
\alias{indKnn}
\title{K-nearest neighbors of nonrespondent units}
\usage{
indKnn(Xr, Xm)
}
\arguments{
\item{Xr}{a matrix without NA values. The rows correspond to respondent units.}

\item{Xm}{a matrix with at least one NA value on each of its rows. The rows correspond to nonrespondent units.}
}
\value{
a matrix containing in columns the nonrespondent units and in rows the nearest respondent units for each nonrespondent in decreasing rank.
}
\description{
Find the k-nearest neighbors of nonrespondent units among respondent units, in terms of Euclidean distance.
The function \code{\link[FNN:knn]{knn}} from the package \code{KNN} is used.
}
\examples{
Xr  <- rbind(c(0.1,0.3,0.4,0.1), c(0.1,0.3,0.2,0.1), c(0.1,0.2,0.3,0.1), c(0.2,0.3,0.2,0.3), c(0.1,0.1,0.2,0.1))
Xm  <- rbind(c(NA,0.1,NA,0.1), c(0.1,NA,0.2,NA))
knn <- indKnn(Xr,Xm)


}
\seealso{
\code{\link[FNN:knn]{FNN::knn}}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
