% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indKnn.R
\name{indKnn}
\alias{indKnn}
\title{K-nearest neighbors of nonrespondent units}
\usage{
indKnn(Xr, Xm)
}
\arguments{
\item{Xr}{a matrix without NA values. Rows correspond to respondent units.}

\item{Xm}{a matrix with some NA values. Rows correspond to nonrespondent units.}
}
\value{
a matrix containing in columns the nonrespondent units and in rows the nearest respondent units for each nonrespondent in decreasing rank.
}
\description{
Find the k-nearest neighbors of nonrespondent units among respondent units, in terms of Euclidean distance.
The function \code{\link[FNN:knn]{knn}} from the package \code{KNN} is used.
}
\examples{


}
\seealso{
\code{\link[FNN:knn]{FNN::knn}}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
