% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissCheeseImput.R
\name{swissCheeseImput}
\alias{swissCheeseImput}
\title{Donor imputation of Swiss cheese nonresponse using the cube method}
\usage{
swissCheeseImput(X, d = NULL, k = NULL, tol = 0.001, max_iter = 2000)
}
\arguments{
\item{X}{a matrix with NA values. Rows correspond to units.}

\item{d}{a vector containing the sampling weights of units. If NULL (default), all sampling weights are equal to 1.}

\item{k}{the number of neighbors considered to impute each nonrespondent. If NULL (default), the smaller k
that makes possible to satisfying calibration equations will be chosen.}

\item{tol}{a tolerance parameter. Default value is 1e-3.}

\item{max_iter}{the maximum number of iterations to consider convergence.}
}
\value{
Returns a list including:

\code{X_new} the imputed matrix of \code{X}.

\code{k} the number of neighbors taken into account.
}
\description{
Impute missing values that have not a particular pattern in a dataset by a donor imputation.
It extends the balanced k-nearest neighbors imputation (Hasler and Tille, 2016) to the treatment of the Swiss cheese nonresponse.
}
\details{
First, the \code{k} nearest neighbors of each nonrespondent are chosen in term of Euclidean distance.
Next, imputation probabilities from nearest neighbors of each nonrespondent  are computed satisfying some calibration constraints for all variables simultaneously
and to sum to one for each nonrespondent. The method of calibration is based on four requirements (see function \link{calibrateKnn} and package vignette
to a complete description of the constraints calibration).
Then, the cube method is used to compute the imputation matrix to choose the donors (see (Deville and Tille, 2004)).
}
\examples{

}
\references{
Deville, J. C. and Tille, Y. (2004). Efficient balanced sampling: the cube method. Biometrika, 91(4), 893-912.
\emph{Biometrika}, 91(4), 893-912

Chauvet, G. and Tille, Y. (2006). A fast algorithm for balanced sampling.
\emph{Computational Statistics}, 21(1), 53-62.
}
\seealso{
\link{indKnn}, \link{calibrateKnn}, \link{cubeImput}, \link{linearImput}
}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
